port.c: Changed return type of GetTickCount() to unsigned int.
"long" is ambiguous on Linux; on 32-bit, it's 32-bit, but on 64-bit,
it's 64-bit. On Windows, it's always 32-bit.

+++ src/gens/port/port.c
@@ -4,12 +4,10 @@
 #include "port.h"
 #include "timer.h"
 
-
-unsigned long GetTickCount()
+unsigned int GetTickCount()
 {
 	return gettime();
 }
-
 
 void SetCurrentDirectory(const char *directory)
 {
